% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boots_var.R
\name{boots_var}
\alias{boots_var}
\title{Bootstrap method}
\usage{
boots_var(data, B)
}
\arguments{
\item{data}{the data given as a vector}

\item{B}{the number of bootstrap samples}
}
\value{
the bootstrap estimate of variance
}
\description{
Compute the bootstrap estimate of variance using R
}
\examples{
\dontrun{
data <- c(1, 2, 0, 4, -3, 7, 2)
boots_var(data = data, B)
}
}
