% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.test.R
\name{skewness.test}
\alias{skewness.test}
\title{Skewness test}
\usage{
skewness.test(data, alpha)
}
\arguments{
\item{data}{the data given as a random variable from N(2, 1) with sample size n=50}

\item{alpha}{the significance level}
}
\value{
the empirical estimation of Type I error
}
\description{
Compute the empirical estimation of Type I error in skewness normality test
}
\examples{
\dontrun{
data <- rnorm(n, 2, 1)
skewness.test(data, alpha)
}
}
