\name{StatComp20057-package}
\alias{StatComp20057-package}
\alias{StatComp20057}
\docType{package}
\title{
  Boostrap method to estimate variance and skewness test of normality
}
\description{
  The fucntion of boostrap method to estimate variance is boots_var(data, B), where the data is your object to analyze, and the B is bootstrap sample size. The fucntion of skewness test of normality is skewness.test(data, alpha), where the data is your object to analyze, and the alpha is significant level.
}
\details{
 (data, B), where the data is your object to analyze, and the B is bootstrap sample size.skewness.test(data, alpha), where the data is your object to analyze, and the alpha is significant level.
}
\author{
Jianping Lan, 1413433806@qq.com

Maintainer: Jianping Lan <1413433806@qq.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
